[prometheus] remove `apps` scrape job
This removes the `apps` scrape job.

The last real use this had was saml-soap-proxy.  In 7658e71f7aff we
removed the rule that matched saml-soap-proxy, but this had the
unintentional consequence of trying to scrape non-hub-app clusters
such as `prometheus` and `static-ingress`.

We should just get rid of the whole job, it's completely redundant.