metadata: deploy fargate task in parallel
deploy a metadata task to the fargate cluster.

lower the cpu value so it fits within fargate constaints.

leave the memory value as is for now so that ec2 based tasks are
unaffected, will remove the container limits in follow up PR.

service discovery is added for consistancy with other services and to
avoid problems if we need it in the future.