scaling: introduce per-app scaling variables
allow scaling out each of the applications seperately.

this was previously complicated as we needed to scale the sizes of the
clusters and then carefully calculate how many tasks we can fit on the
boxes... no such complexity exists with fargate and we can simply ask
for N-number of each application to scale-out as we see fit

defaults are set to 2 for all apps, which is the minium to ensure
zero-downtime rolling deployment of tasks.