BAU: Remove Splunk vars from saml-engine
These vars were being provided to saml-engine so that our fork of the
splunk java-logging library could consume them and ship logs to Splunk.
We were using the appender on one particular class. This was set up when
all the application logs went to Kibana.

As all the app logs now go to Splunk anyway, we no longer need the
special use case with the library and can drop it from the hub.