HUB-733 Terraform updates to specify SENTRY_ENV
Sentry now provides a mechnism for reporting environments.  But
environments to sentry are different to the way some of our apps run.
e.g. Rails apps always run in a production environment which is reported
to sentry.  Sentry then sets the environment to production even if the
app in question is running in our staging or integration environments.