HUB-738: Refactor metadata to separate module
This creates a new module, metadata. The point of this is to allow the
metadata ECS service to be applied separately from the hub terraform -
the ultimate goal being a separate build/deploy pipeline for the
metadata rather than being part of the hub pipeline.

This also adds some config to the hub module to allow the metadata
config to be "turned off" on an environment basis. This is to allow
migration of the metadata to the new pipeline by environment. The
default specified is maintain the current behaviour, making these
changes a no-op.

A few more outputs have also been specified in the hub module - this so
they can be picked up by the new metadata module via the `remote_state`
data.

There will be a separate PR on the verify-infrastructure-config repo
that will invoke the new metadata module as well as configure the
current hub module to stop managing the metadata infrastructure per
environment.

The new module will create a new metadata service which is identical to
the existing service managed by the hub terraform module. This service
is placed in the same load balancer group as the original service. At
this point both services will be receiving traffic. We can then update
the config (set `manage_metadata` to false) to destroy the original
service.

This also upgrades Terraform to v0.13. The refactor of the metadata
terraform out of the hub terraform uses a feature not available in v0.12
(`count`). Upgrading has meant a few changes to other syntax issues are
required, which are addressed here.