HUB-748 Remove hosted instance of Sentry
Now that we use the cloud version of Sentry we don't need to host our own version.  Sending `staging` events to the cloud version is more cost effective than running an ec2 instance and database etc.

This should remove anything that is used solely by the hoseted instance.  Variables remain for use by with the cloud instance.