HUB-957: Increase matomo's php-fpm max_children
We were previously just vendoring this Matomo base image. This allows us
to build it and slip in an extra config file.

This `z` prefix on the config file is important as the config files in
the directory are read alphabetically and we need this to overwrite the
values set in `www.conf`.

The new config file ups our max_children from 5 (the default) to 20.
This is being done due to issues with process starvation when the Matomo
GUI was sending long running requests to the backend.

`start_servers` is set to 20 too so we don't have to warm up if a new
Matomo task starts up during a busy period.

`max_spare_servers` can't be less than `start_servers`.

If the process are left idle for a while php-fpm should kill them.