BAU: Set nginx-tls image /etc/hosts owner in entrypoint
Since we started building the nginx-tls image in the pipeline with the
vito/oci-build-task image, the `build-nginx-tls` Concourse job has been
failing.

The build attempts to change the owner of the `/etc/hosts` directory to
`nginx`. This fails with the error message:

        `chown: /etc/hosts: Read-only file system`

The oci-build-task uses BuildKit directly under the hood, rather than
with Docker. I believe this is causing the issue.

It seems that `/etc/hosts` is treated specially by the build runtime,
and BuildKit does not allow writes to the dir. There is some discussion
of it in [this issue.](https://github.com/moby/buildkit/issues/1267).

It seems that [you need to manually turn BuildKit on in Docker](https://docs.docker.com/develop/develop-images/build_enhancements/), which
would explain why this has only just broken.

To get around this issue, we can change the owner of `/etc/hosts` after
the build, in the entrypoint script. To do this we need to be the root
user, which is why the `USER nginx` instruction has been removed from the
Dockerfile.

Once the owner of `/etc/hosts` has been changed we can switch to the
`nginx` user. To do this from the entrypoint script the nginx users
needs to have a login shell set. That's what the `sed` command is doing.

I'm not 100% sure that this is the right approach for this. I'm not sure
why the `/etc/hosts` dir needs to be owned by nginx in the first place,
but it's probably been done for a good reason. But the simplest fix
would be to not change the ownership at all.