BAU: Increase memory available to Matomo PHP
We've seen a few errors in the archiving logs and the docs seems to
suggest they may be caused by the server running out of memory.

We currently have the memory limit available to php set to 128MB. This
bumps it to 1G.

It's worth noting that this limit is per process, not an overall limit.
Having said that we consistently use just 2% of available memory so
even a nearly ten fold increase in use should be manageable.