BAU: Correctly set php memory limit for matomo
The last attempt to update the memory limit failed as the php.ini file
was never loaded.

This creates a new custom ini file with just the new memory limit and
adds it to the directory that php scans for ini files.

I've tested this locally and the command `php -i | grep memory_limit`
shows that the new value has been loaded.