HUB-765: Alert on metadata signing certificate four week expiry
The policy at the moment is to alert to 2ndline if certificates have 2
weeks until their expiry.

We need to alert sooner for the metadata signing certificate as we would
never want to sign metadata with a certificate that has a shorter life
than the metadata itself.

This adds a new panel with an attached alert that will fire at the four
week mark. Unfortunately you can't add more than one alert to a panel,
so I've had to create a new one. It's not for looking at, just for
having an alert attached.